requirements.txt
torch==1.7.0
torchvision==0.8.1
Pillow==8.0.1
opencv-python==3.4.0.14
tqdm==4.54.0